+++ src/NppPlugin.Host/PluginInfrastructure/Scintilla_iface.cs
@@ -21,7 +21,7 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
         /// <summary>
         /// environment specific window handle/pointer
         /// </summary>
-        public IntPtr hwndFrom; 
+        public IntPtr hwndFrom;
 
         /// <summary>
         /// CtrlID of the window issuing the notification
@@ -31,7 +31,7 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
         /// <summary>
         /// The SCN_* notification Code
         /// </summary>
-        public uint Code;       
+        public uint Code;
     }
 
     [StructLayout(LayoutKind.Sequential)]
@@ -73,7 +73,7 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
         /// Character of the notification - eg keydown
         /// SCN_CHARADDED, SCN_KEY, SCN_AUTOCCOMPLETE, SCN_AUTOCSELECTION, SCN_USERLISTSELECTION
         /// </summary>
-        public char Character { get { return (char) character; } }
+        public char Character { get { return (char)character; } }
     }
 
     [Flags]
@@ -2991,7 +2991,7 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
         /// <param name="cpmin">range to search</param>
         /// <param name="cpmax">range to search</param>
         /// <param name="searchText">the search pattern</param>
-        public TextToFind(int cpmin, int cpmax, string searchText)
+        public TextToFind(IntPtr cpmin, IntPtr cpmax, string searchText)
         {
             _sciTextToFind.chrg.cpMin = cpmin;
             _sciTextToFind.chrg.cpMax = cpmax;
